feat: Implement Google login functionality and refactor user payload generation
